requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
dj-database-url==2.0.0
Django==4.1
django-cleanup==7.0.0
django-embed-video==1.4.8
django-on-heroku==1.1.2
idna==3.4
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
psycopg2-binary==2.9.6
regex==2023.5.5
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
whitenoise==6.4.0